Move the image cache card to a new Media settings tab
The image cache card was placed on the Configurations tab, which is for KIP
configuration management (backup, restore, import, export). Move it to a new
Media tab, placed before Configurations, so the image cache lives with media
concerns rather than config management.

Extracts the image-cache logic into a SettingsMediaComponent (with its own
spec) and removes it from SettingsConfigComponent. Updates the Image widget
help doc to point at Settings → Media → Image Cache.
