+++ src/app/core/components/options/configuration/config.component.scss
@@ -27,26 +27,6 @@ h3 {
   margin-top: 5px;
 }
 
-.image-cache-size {
-  display: flex;
-  align-items: baseline;
-  gap: 8px;
-  margin-bottom: 8px;
-}
-
-.image-cache-label {
-  opacity: 0.7;
-}
-
-.image-cache-value {
-  font-weight: 600;
-}
-
-// Horizontal gap between the Refresh and Purge buttons (footer is text-align:end, not flex).
-.image-cache-actions button + button {
-  margin-left: 8px;
-}
-
 a {
   font-size: 14px;
 }
