+++ src/app/core/components/options/media/media.component.html
@@ -1,0 +1,35 @@
+<div class="page-content">
+  <div class="flex-container">
+    <div class="flex-item-rounded-card rounded-card-color">
+      <h3>Image Cache</h3>
+      <p style="margin-bottom: 16px">
+        Uploaded images are stored on the Signal K server and resized copies are
+        cached on disk. Purging removes the generated copies; originals are kept
+        and copies regenerate on demand.
+      </p>
+      <div class="image-cache-size">
+        <span class="image-cache-label">On-disk cache:</span>
+        <span class="image-cache-value">{{ imageCacheDisplay() }}</span>
+      </div>
+      <div class="formActionFooter image-cache-actions">
+        <mat-divider class="formActionDivider"></mat-divider>
+        <button
+          mat-stroked-button
+          type="button"
+          (click)="refreshImageCache()"
+          >
+          Refresh
+        </button>
+        <button
+          mat-flat-button
+          type="button"
+          color="accent"
+          [disabled]="imageCachePurging()"
+          (click)="purgeImageCache()"
+          >
+          Purge
+        </button>
+      </div>
+    </div>
+  </div>
+</div>
