+++ src/app/core/components/options/media/media.component.scss
@@ -1,0 +1,36 @@
+:host {
+  display: block;
+  height: 100%;
+  width: 100%;
+}
+
+.page-content {
+  width: 100%;
+  overflow-y: auto;
+  scroll-behavior: smooth;
+  padding: 0px 0px 10px 0px;
+}
+
+h3 {
+  margin-top: 5px;
+}
+
+.image-cache-size {
+  display: flex;
+  align-items: baseline;
+  gap: 8px;
+  margin-bottom: 8px;
+}
+
+.image-cache-label {
+  opacity: 0.7;
+}
+
+.image-cache-value {
+  font-weight: 600;
+}
+
+// Horizontal gap between the Refresh and Purge buttons (footer is text-align:end, not flex).
+.image-cache-actions button + button {
+  margin-left: 8px;
+}
