+++ src/app/core/components/options/tabs/tabs.component.html
@@ -12,6 +12,9 @@
   <mat-tab label="Units">
       <settings-units class="tab-content" />
   </mat-tab>
+  <mat-tab label="Media">
+      <settings-media class="tab-content" />
+  </mat-tab>
   <mat-tab label="Configurations">
       <settings-config class="tab-content" />
   </mat-tab>
