+++ src/app/core/components/options/tabs/tabs.component.ts
@@ -6,6 +6,7 @@ import { SettingsDisplayComponent } from '../display/display.component';
 import { MatTabGroup, MatTab } from '@angular/material/tabs';
 import { PageHeaderComponent } from '../../page-header/page-header.component';
 import { SettingsConfigComponent } from '../configuration/config.component';
+import { SettingsMediaComponent } from '../media/media.component';
 
 @Component({
     selector: 'tabs',
@@ -19,6 +20,7 @@ import { SettingsConfigComponent } from '../configuration/config.component';
       SettingsUnitsComponent,
       SettingsDisplayComponent,
       SettingsNotificationsComponent,
+      SettingsMediaComponent,
       SettingsConfigComponent,
       PageHeaderComponent
   ]
